This fixes a race condition between the calling thread and the SocketThread.
It was possible that once the request had been written, the response was coming back before the "this.asyncResult = null" (line 258) statement was being called. This was causing a valid response to be lost and "null" being returned. This fix first checks to ensure no response has been received yet, and only then to wait for a second at most for a response.